add use-item lightgcn script

+++ helper/build_train_dl.py
@@ -76,7 +76,7 @@ def _build_gnn_train_dl(config, data):
         # build node_collate_graph
         g_src = io.load_pickle(osp.join(data_root, 'train_undi_csr_src_indices.pkl'))
         g_dst = io.load_pickle(osp.join(data_root, 'train_undi_csr_indices.pkl'))
-        node_collate_graph = dgl.graph((g_src, g_dst), idtype=torch.int32)
+        node_collate_graph = dgl.graph((g_src, g_dst))
 
         train_dl = EdgeBased_Sampling_Block_TrainDataLoader(
             info, E_src, E_dst,
